Polish 'juniper' crate changes, vol.1
