+++ juniper/src/macros/mod.rs
@@ -1,6 +1,6 @@
 //! Helper traits for macros.
 
+pub mod helper;
+
 #[cfg(test)]
 mod tests;
-
-pub mod subscription_helpers;
