+++ juniper/src/types/base.rs
@@ -7,7 +7,6 @@ use crate::{
     schema::meta::{Argument, MetaType},
     value::{DefaultScalarValue, Object, ScalarValue, Value},
     GraphQLEnum,
-    types::async_await::DynGraphQLValueAsync,
 };
 
 /// GraphQL type kind
@@ -291,19 +290,11 @@ where
 
 crate::sa::assert_obj_safe!(GraphQLValue<Context = (), TypeInfo = ()>);
 
+/// Helper alias for naming [trait objects][1] of [`GraphQLValue`].
+///
+/// [1]: https://doc.rust-lang.org/reference/types/trait-object.html
 pub type DynGraphQLValue<S, C, TI> =
     dyn GraphQLValue<S, Context = C, TypeInfo = TI> + Send + Sync + 'static;
-
-pub trait AsDynGraphQLValue<S: ScalarValue = DefaultScalarValue> {
-    type Context;
-    type TypeInfo;
-
-    fn as_dyn_graphql_value(&self) -> &DynGraphQLValue<S, Self::Context, Self::TypeInfo>;
-
-    fn as_dyn_graphql_value_async(&self) -> &DynGraphQLValueAsync<S, Self::Context, Self::TypeInfo>;
-}
-
-crate::sa::assert_obj_safe!(AsDynGraphQLValue<Context = (), TypeInfo = ()>);
 
 /// Primary trait used to expose Rust types in a GraphQL schema.
 ///
